Refactor code structure and improve readability
- Updated function signatures to use consistent naming conventions (e.g., `code_conv` to `code_conv_t`).
- Enhanced memory management in the first pass by dynamically allocating and reallocating data structures.
- Improved error handling and reporting for label declarations and data processing.
- Refactored utility functions for better clarity and consistency in naming.
- Added base-4 address representation in output files for better readability.
- Streamlined the process of handling data directives and external labels.
- Fixed potential memory leaks by ensuring proper deallocation of dynamically allocated memory.
- Updated the memory map printing function to correctly format output.
- Added new output files for external references.

+++ src/memory_map/memory_map.c
@@ -44,7 +44,7 @@ void memory_map_print(const MemoryMap *map) {
     printf("Memory Map (used cells):\n");
     for (i = 0; i < map->size; ++i) {
         if (map->cells[i].used) {
-            printf("%03zu: %04x\n", i, map->cells[i].value);
+            printf("%03lu: %04x\n", (unsigned long)i, map->cells[i].value);
         }
     }
 }
